+++ source/code/plugin/kubelet_utils.rb
@@ -19,175 +19,5 @@ class KubeletUtils
                 return [cpu_capacity, memory_capacity]
             end
         end
-
-        def getContainerInventoryRecords(batchTime, clusterCollectEnvironmentVar, @containerCGroupCache)
-            containerInventoryRecords = Array.new
-            begin
-                response = CAdvisorMetricsAPIClient.getPodsFromCAdvisor(winNode: nil)
-                if !response.nil? && !response.body.nil?
-                    podList = JSON.parse(response.body)
-                    if !podList.nil? && !podList.empty? && podList.key?("items") && !podList["items"].nil? && !podList["items"].empty?
-                        podList["items"].each do |item|
-                            containersInfoMap = getContainersInfoMap(item)
-                            containerInventoryRecord = {}
-                            if !item["status"].nil? && !item["status"].empty?
-                                if !item["status"]["containerStatuses"].nil? && !item["status"]["containerStatuses"].empty?
-                                    item["status"]["containerStatuses"].each do |containerStatus|
-                                     containerInventoryRecord["CollectionTime"] = batchTime #This is the time that is mapped to become TimeGenerated
-                                     containerName = containerStatus["name"]
-                                     containerId = containerStatus["containerID"].split('//')[1]
-                                     containerInventoryRecord["InstanceID"] = containerId
-                                    # imagedId is of the format - repo@sha256:imageid
-                                    imageIdValue =  containerStatus["imageID"]
-                                    if !imageIdValue.nil? && !imageIdValue.empty?
-                                        atLocation = imageIdValue.index("@")
-                                        if !atLocation.nil?
-                                          containerInventoryRecord["ImageId"] = imageIdValue[(atLocation + 1)..-1]
-                                        end
-                                    end
-                                    # image is of the format - repository/image:imagetag
-                                    imageValue = containerStatus["image"]
-                                    if !imageValue.nil? && !imageValue.empty?
-                                        # Find delimiters in the string of format repository/image:imagetag
-                                        slashLocation = imageValue.index("/")
-                                        colonLocation = imageValue.index(":")
-                                        if !colonLocation.nil?
-                                          if slashLocation.nil?
-                                            # image:imagetag
-                                            containerInventoryRecord["Image"] = imageValue[0..(colonLocation - 1)]
-                                          else
-                                            # repository/image:imagetag
-                                            containerInventoryRecord["Repository"] = imageValue[0..(slashLocation - 1)]
-                                            containerInventoryRecord["Image"] = imageValue[(slashLocation + 1)..(colonLocation - 1)]
-                                          end
-                                          containerInventoryRecord["ImageTag"] = imageValue[(colonLocation + 1)..-1]
-                                        end
-                                    elsif !imageIdValue.nil? && !imageIdValue.empty?
-                                        # Getting repo information from imageIdValue when no tag in ImageId
-                                        if !atLocation.nil?
-                                           containerInventoryRecord["Repository"] = imageIdValue[0..(atLocation - 1)]
-                                        end
-                                    end
-                                     containerInventoryRecord["ExitCode"] = 0
-                                     if !containerStatus["state"].nil? && !containerStatus["state"].empty?
-                                        containerState = containerStatus["state"]
-                                        if containerState.key?("running")
-                                            containerInventoryRecord["State"] = "Running"
-                                            containerInventoryRecord["StartedTime"] = containerState["running"]["startedAt"]
-                                        elsif containerState.key?("terminated")
-                                            containerInventoryRecord["State"] = "Terminated"
-                                            containerInventoryRecord["StartedTime"] = containerState["terminated"]["startedAt"]
-                                            containerInventoryRecord["FinishedTime"] = containerState["terminated"]["finishedAt"]
-                                            containerInventoryRecord["ExitCode"] = containerState["terminated"]["exitCode"]
-                                        elsif containerState.key?("waiting")
-                                            containerInventoryRecord["State"] = "Waiting"
-                                        end
-                                     end
-                                     containerInfoMap = containersInfoMap[containerName]
-                                     containerInventoryRecord["ElementName"] = containerInfoMap["ElementName"]
-                                     containerInventoryRecord["Computer"] = containerInfoMap["Computer"]
-                                     containerInventoryRecord["ContainerHostname"] = containerInfoMap["ContainerHostname"]
-                                     containerInventoryRecord["CreatedTime"] = containerInfoMap["CreatedTime"]
-                                     containerInventoryRecord["EnvironmentVar"] = containerInfoMap["EnvironmentVar"]
-                                     containerInventoryRecord["Ports"] = containerInfoMap["Ports"]
-                                     containerInventoryRecord["Command"] = containerInfoMap["Command"]
-                                     if !clusterCollectEnvironmentVar.nil? && !clusterCollectEnvironmentVar.empty? && clusterCollectEnvironmentVar.casecmp("false") == 0
-                                        containerInventoryRecord["EnvironmentVar"] = ["AZMON_CLUSTER_COLLECT_ENV_VAR=FALSE"]
-                                      else
-                                        containerInventoryRecord["EnvironmentVar"]  = KubeletUtils.obtainContainerEnvironmentVars(containerId, @containerCGroupCache)
-                                     end          
-                                     containerInventoryRecords.push containerInventoryRecord
-                                    end
-                                end
-                            end
-                        end
-                    end
-                end
-            rescue => error
-                @Log.warn("kubelet_utils::getContainerInventoryRecords : Get Container Inventory Records failed: #{error}")
-            end
-            return containerInventoryRecords
-        end
-
-        def getContainersInfoMap(item)
-            containersInfoMap = {}
-            begin
-                nodeName = (!item["spec"]["nodeName"].nil?) ? item["spec"]["nodeName"] : ""
-                createdTime = item["metadata"]["creationTimestamp"]
-                if !item.nil? && !item.empty? && item.key?("spec") && !item["spec"].nil? && !item["spec"].empty?
-                    if !item["spec"]["containers"].nil? && !item["spec"]["containers"].empty?
-                        item["spec"]["containers"].each do |container|
-                            containerInfoMap = {}
-                            containerName = container["name"]
-                            containerInfoMap["ElementName"] = containerName
-                            containerInfoMap["Computer"] = nodeName
-                            containerInfoMap["ContainerHostname"] = nodeName
-                            containerInfoMap["CreatedTime"] = createdTime                           
-                            portsValue = container["ports"]
-                            portsValueString = (portsValue.nil?) ? "" : portsValue.to_s
-                            containerInfoMap["Ports"] = portsValueString
-                            cmdValue = container["command"]
-                            cmdValueString = (cmdValue.nil?) ? "" : cmdValue.to_s
-                            containerInfoMap["Command"] = cmdValueString
-
-                            containersInfoMap[containerName] = containerInfoMap
-                        end
-                    end
-                end
-            rescue => error
-                @Log.warn("kubelet_utils::getContainersInfoMap : Get Container Info Maps failed: #{error}")
-            end
-            return containersInfoMap
-        end
-
-        def obtainContainerEnvironmentVars(containerId, containerCGroupCache)
-            envValueString = ""
-            begin                                
-                unless containerCGroupCache.has_key?(containerId)
-                    Dir["/hostfs/proc/*/cgroup"].each do| filename|
-                        if  File.file?(filename) && File.foreach(filename).grep(/#{containerId}/).any?
-                            # file full path is /hostfs/proc/<cGroupPid>/cgroup
-                            cGroupPid = filename.split("/")[3]
-                            if containerCGroupCache.has_key?(containerId)
-                                temCGroupPid = containerCgroupCache[containerId]
-                                if temCGroupPid > cGroupPid
-                                   containerCgroupCache[containerId] = cGroupPid
-                                end
-                            else
-                                containerCGroupCache[containerId] = cGroupPid
-                            end
-                        end
-                    end
-                end
-                cGroupPid = containerCGroupCache[containerId]
-                if !cGroupPid.nil?
-                    environFilePath = "/hostfs/proc/#{cGroupPid}/environ"
-                    if File.exist?(environFilePath)
-                        # Skip environment variable processing if it contains the flag AZMON_COLLECT_ENV=FALSE
-                        # Check to see if the environment variable collection is disabled for this container.
-                        if File.foreach(environFilePath).grep(/AZMON_COLLECT_ENV=FALSE/i).any?
-                            envValueString = ["AZMON_COLLECT_ENV=FALSE"]
-                            $log.warn("Environment Variable collection for container: #{containerName} skipped because AZMON_COLLECT_ENV is set to false")
-                        else
-                            fileSize = File.size(environFilePath)
-                            # Restricting the ENV string value to 200kb since the size of this string can go very high
-                            envVars = File.read(environFilePath, 200000).split(" ")
-                            envValueString = envVars.to_s                        
-                            if fileSize > 200000                               
-                                lastIndex = envValueString.rindex("\", ")
-                                if !lastIndex.nil?
-                                   envValueStringTruncated = envValueString.slice(0..lastIndex) + "]"
-                                   envValueString = envValueStringTruncated                              
-                                end                              
-                            end                         
-                        end
-                    end
-                end
-            rescue => error
-                 @Log.warn("kubelet_utils::obtainContainerEnvironmentVars : obtain Container Environment vars failed: #{error} for containerId: #{containerId}")
-            end
-            return envValueString
-         end
-
     end
 end
